Split up catalua_bindings.cpp into 3 files
